Implementar comando comprar, corrigir loja

+++ ZaynBot/RPG/Entidades/RPGItem.cs
@@ -13,7 +13,7 @@ namespace ZaynBot.RPG.Entidades
         public TipoItemEnum TipoItem { get; set; }
         public TipoExpEnum TipoExp { get; set; }
         public int DurabilidadeMax { get; set; }
-        public double PrecoCompra { get; set; }
+        public int PrecoCompra { get; set; }
         public double PrecoVenda { get; set; }
 
         public double AtaqueFisico { get; set; }
@@ -23,7 +23,7 @@ namespace ZaynBot.RPG.Entidades
         public double MagiaRestaura { get; set; }
         public double FomeRestaura { get; set; }
 
-        public RPGItem(int id, string nome, TipoItemEnum tipo, double preco = 1)
+        public RPGItem(int id, string nome, TipoItemEnum tipo, int preco = 1)
         {
             Id = id;
             Nome = nome;
@@ -33,7 +33,7 @@ namespace ZaynBot.RPG.Entidades
             TipoExp = TipoExpEnum.Nenhum;
         }
 
-        public RPGItem(int id, string nome, TipoItemEnum tipo, int durabilidade, double preco, TipoExpEnum tipoExp) : this(id, nome, tipo, preco)
+        public RPGItem(int id, string nome, TipoItemEnum tipo, int durabilidade, int preco, TipoExpEnum tipoExp) : this(id, nome, tipo, preco)
         {
             TipoExp = tipoExp;
             DurabilidadeMax = durabilidade;
